requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.31
gunicorn==22.0.0
Flask-Caching==2.3.1
pydantic==2.8.2
psycopg2-binary==2.9.9
blinker==1.9.0

